move AWS::DynamoDB::Table to models
